tests/wpt/tests/css/css-transforms/parsing/perspective-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transform Module Level 2: parsing perspective with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#perspective-property">
<meta name="assert" content="perspective supports only the grammar 'none | <length [0,∞]>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("perspective", "1000");
test_invalid_value("perspective", "-1px");
test_invalid_value("perspective", "80%");
</script>
</body>
</html>
